# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sixpeng API
  version: 0.1.0
  description: |
    AI 기반 Borderless Teamwork 플랫폼 API.

    이 문서는 실제 구현된 엔드포인트를 기술한다.
    현재 범위: `accounts`, `projects`, `translation`, `messages`, `integrations`.
    `briefings` / `qna` 는 아직 미구현이라 스펙만 정의되어 있다.

    ## 인증
    Spring Security 세션 기반이다. 로그인이 끝나면 `JSESSIONID` 쿠키가 내려가고,
    이후 모든 요청은 이 쿠키로 인증된다.

    쿠키 속성: `HttpOnly; Secure; SameSite=None`
    → 프론트는 반드시 `fetch(..., { credentials: 'include' })` 로 호출해야 한다.

    ## CSRF
    `SameSite=None` 이라 SameSite 방어가 없으므로 CSRF 토큰이 켜져 있다.
    - 서버가 `XSRF-TOKEN` 쿠키를 내려준다 (HttpOnly 아님 → JS 로 읽을 수 있음).
    - **상태를 바꾸는 요청(POST/PATCH/PUT/DELETE)** 은 그 값을 `X-XSRF-TOKEN` 헤더로 되돌려 보내야 한다.
    - GET 은 토큰이 필요 없다.
    - 토큰 없이 보내면 `403`.

    ## CORS
    `app.cors.allowed-origins` 에 등록된 오리진만 허용하며 `Access-Control-Allow-Credentials: true` 로 응답한다.
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
paths:
...
  /projects:
    get:
      tags: [projects]
      summary: 내가 참여한 프로젝트 목록
      description: |
        최근 생성 순(id 내림차순).

        `status` 를 주면 해당 상태만, 없으면 상태와 무관하게 전체를 반환한다.
        어느 쪽이든 내가 참여 중인 프로젝트로 범위가 제한된다.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProjectStatus"
          description: 생략하면 전체. enum 에 없는 값이면 400.
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [projects]
      summary: 프로젝트 생성
      description: |
        생성자는 `role=PM` 으로 참여자에 자동 등록된다.
        요구사항 5-5 에 따라 **AI 팀원도 `role=AI` 로 함께 자동 합류**하므로,
        생성 직후 참여자 목록은 항상 2명(생성자 + AI 팀원)이다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CsrfForbidden"
  /projects/{projectId}:
    get:
      tags: [projects]
      summary: 프로젝트 상세
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotParticipant"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
  /projects/{projectId}/status:
    patch:
      tags: [projects]
      summary: 프로젝트 상태 변경
      description: |
        참여자면 누구나 바꿀 수 있다 (초대/내보내기와 같은 규칙).
        상태 전이에 제약은 없다 — 완료된 프로젝트를 다시 진행중으로 되돌릴 수 있다.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectStatusUpdateRequest"
      responses:
        "200":
          description: 변경 성공. 변경된 프로젝트 전체를 반환한다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotParticipant"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
  /projects/{projectId}/members:
    get:
      tags: [projects]
      summary: 프로젝트 참여자 목록
      description: |
        사람 팀원과 AI 팀원이 함께 내려온다.
        AI 팀원 판별은 `isAiTeammate` 플래그로 한다 (`memberId` 는 환경마다 달라지므로 하드코딩 금지).
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotParticipant"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
    post:
      tags: [projects]
      summary: 참여자 초대
      description: |
        **이미 Google 로그인으로 가입한 회원만** 초대할 수 있다(초대장 테이블이 없음).
        가입하지 않은 이메일이면 `404 M001`.
        응답 본문은 없다.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMemberInviteRequest"
      responses:
        "201":
          description: 초대 성공 (본문 없음)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotParticipant"
        "404":
          description: 프로젝트 또는 초대 대상 회원 없음 (`P001` / `M001`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 이미 참여 중인 회원 (`P004`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/members/{memberId}:
    delete:
      tags: [projects]
      summary: 참여자 내보내기
      description: |
        프로젝트 생성자는 내보낼 수 없다 (`400 P005`).
        AI 팀원도 모든 프로젝트에 항상 참여해 있어야 하므로 내보낼 수 없다 (`400 P006`).
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 삭제 성공
        "400":
          description: 생성자 또는 AI 팀원은 제거할 수 없음 (`P005` / `P006`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotParticipant"
        "404":
          description: 프로젝트 또는 참여자 없음 (`P001` / `P003`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Unauthorized:
      description: 로그인하지 않음 (`C006`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: 입력값이 잘못됨 (`C001`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    CsrfForbidden:
      description: CSRF 토큰 누락/불일치 (`C005`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotParticipant:
      description: 해당 프로젝트의 참여자가 아님 (`P002`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ProjectNotFound:
      description: 프로젝트 없음 (`P001`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ProjectStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED]
      description: |
        프로젝트 진행 상태. 생성 시 `PENDING`.

        | 값 | 표시 |
        | --- | --- |
        | `PENDING` | 진행전 |
        | `IN_PROGRESS` | 진행중 |
        | `COMPLETED` | 완료 |

        표시 문구는 프론트가 매핑한다. 이 서비스는 번역이 핵심 기능이라
        한글 문구를 응답에 담지 않는다.
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        createdBy:
          type: integer
          format: int64
          description: 생성한 회원의 id
        createdAt:
          type: string
          format: date-time
      example:
        id: 1
        name: sixpeng
        status: PENDING
        createdBy: 1
        createdAt: "2026-08-09T22:51:07.260082"
    ProjectCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
      description: |
        `status` 는 받지 않는다. 새 프로젝트는 항상 `PENDING` 으로 만들어지며,
        이후 `PATCH /projects/{projectId}/status` 로 바꾼다.
    ProjectStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ProjectStatus"
    ProjectMember:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        name:
          type: string
        role:
          type: string
          nullable: true
          description: PM, Backend, Design 등
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        isAiTeammate:
          type: boolean
          description: |
            AI 팀원이면 true. **AI 팀원 판별은 반드시 이 필드로 한다.**
            `memberId` 는 환경(로컬/개발/운영)마다 값이 달라지므로 하드코딩하면 안 된다.
      example:
        memberId: 2
        name: AI 팀원
        role: AI
        country: null
        timezone: null
        isAiTeammate: true
    ProjectMemberInviteRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: 이미 가입한 회원의 이메일
        role:
          type: string
          maxLength: 255
    ErrorResponse:
      type: object
      description: 모든 에러 응답의 공통 형태
      properties:
        code:
          type: string
          description: |
            C001 잘못된 입력 / C005 접근 거부 (CSRF 포함) / C006 미인증
            M001 회원 없음 / M002 AI 팀원 계정 미초기화 / M003 번역 표시 토글이 둘 다 false
            P001 프로젝트 없음 / P002 참여자 아님 / P003 참여자 없음
            P004 이미 참여 중 / P005 생성자는 제거 불가 / P006 AI 팀원은 제거 불가
            T001 번역 API 실패
            I001 Notion OAuth 인증 실패 / I002 잘못된 OAuth state
        message:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
      example:
        code: P002
        message: Not a participant of the project
        errors: []
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        value:
          type: string
        reason:
          type: string
...
security:
  - sessionCookie: []